Completed datapath implementation to compute angle offset, alpha and abs_beta - angle offset is computed over 2 cycles. angle_offset = column_count * 0.375 - alpha and beta are computed over 3 cycles. alpha = angle + 30 - angle_offset, computed in cycle 0 (S_FIND_ALPHA_BETA_0) - beta = angle_offset - 30, computed in cycle 1 (S_FIND_ALPHA_BETA_1) - abs_beta = abs(beta) computed in cycle 2 (S_FIND_ALPHA_BETA_2)
alpha is the current angle at which a ray is being cast. To get it, we
shift to the left of the FOV from the player angle (angle + 30) and then
subtract in increments of 0.375 degrees till we span the entire FOV.

beta is the angle of the ray relative to player angle, used to reverse
the fishbowl effect. absolute value taken since cos(beta) = cos(-beta)
and the lookup table does not contain negative values

- Included simulation file for find_slice_size that demonstrates these
datapath calculations